Fix bug generated by @hoshinotented

+++ core/src/org/ice1000/jimgui/JImGuiFont.java
@@ -16,7 +16,7 @@ public class JImGuiFont extends JImGuiFontGen {
 	}
 
 	public void setFallbackChar(char wChar) {
-        setFallbackChar(wChar);
+		setFallbackChar((short) wChar);
 	}
 
 	public native @MagicConstant(flagsFromClass = JImFontAtlasFlags.class)
